+++ src/pages/campaign-url-builder/_GeneratedURL/_index.tsx
@@ -1,9 +1,9 @@
 import * as React from "react"
-import useStyles from "./_useStyles"
-import { usePersistentBoolean } from "../../hooks"
-import { StorageKey } from "../../constants"
+import useStyles from "../_useStyles"
+import { usePersistentBoolean } from "../../../hooks"
+import { StorageKey } from "../../../constants"
 import useShortenLink from "./_useShortenLink"
-import { websiteUrlFor } from "./_params"
+import { websiteUrlFor } from "../_params"
 import Paper from "@material-ui/core/Paper"
 import { Error as ErrorIcon } from "@material-ui/icons"
 import WarningsFor from "./_WarningsFor"
@@ -11,7 +11,7 @@ import Typography from "@material-ui/core/Typography"
 import TextField from "@material-ui/core/TextField"
 import FormControlLabel from "@material-ui/core/FormControlLabel"
 import Checkbox from "@material-ui/core/Checkbox"
-import CopyButton from "../../components/CopyButton"
+import CopyButton from "../../../components/CopyButton"
 import Button from "@material-ui/core/Button"
 
 interface GeneratedUrlProps {
